+++ source/SAMRAI/pdat/EdgeData.h
@@ -323,7 +323,7 @@ public:
     *
     * @pre dynamic_cast<const EdgeOverlap *>(&overlap) != 0
     */
-   virtual int
+   virtual size_t
    getDataStreamSize(
       const hier::BoxOverlap& overlap) const;
